refactors files for better modularity
